Refactor test formatting and minor code cleanups
- Reformatted all test methods in ChannelQueueTests.cs for consistent indentation; no logic changes.
- Updated .editorconfig "last updated" date.
- Removed unused using statements in DiagnosticEntryTests.cs and PersonComparerTests.cs.
- Fixed type assertion in ExceptionExtensionsTests.cs for IsLogged metadata.
- Used discard for unused out variable in InformationAttributeDocGeneratorTests.cs.
- Simplified assertions and removed unnecessary casts in EnumerableExtensionsTests.cs and ListExtensionsTests.cs.
- Used 'var' for result variable in PersonComparerByLastNameTests.cs.

+++ source/Unit Tests/dotNetTips.Spargine.Extensions.Tests/EnumerableExtensionsTests.cs
@@ -1256,7 +1256,7 @@ public class EnumerableExtensionsTests
 
 		var result = numbers.FastLongCount(n => n > 0);
 
-		Assert.AreEqual((long)largeCount, result);
+		Assert.AreEqual(largeCount, result);
 		Assert.IsInstanceOfType<long>(result);
 	}
 
@@ -1318,7 +1318,7 @@ public class EnumerableExtensionsTests
 		var expectedCount = people.Count(p => p.FirstName.Contains('a', StringComparison.OrdinalIgnoreCase));
 		var result = people.FastLongCount(p => p.FirstName.Contains('a', StringComparison.OrdinalIgnoreCase));
 
-		Assert.AreEqual((long)expectedCount, result);
+		Assert.AreEqual(expectedCount, result);
 	}
 
 	[TestMethod]
